+++ common/src/main/java/net/morimori0317/gamemenumodoption/ClientConfig.java
@@ -1,0 +1,11 @@
+package net.morimori0317.gamemenumodoption;
+
+public interface ClientConfig {
+    boolean showNotificationModUpdate();
+
+    boolean renderModListBackground();
+
+    boolean modMenuStyle();
+
+    boolean overwriteModButtonText();
+}