+++ common/src/main/java/net/morimori0317/gamemenumodoption/GMMOPauseScreen.java
@@ -1,0 +1,9 @@
+package net.morimori0317.gamemenumodoption;
+
+import net.minecraft.client.gui.components.Button;
+
+public interface GMMOPauseScreen {
+    Button getModOptionButton();
+
+    boolean isShowPauseMenu();
+}