+++ common/src/main/java/net/morimori0317/gamemenumodoption/GameMenuModOption.java
@@ -1,0 +1,11 @@
+package net.morimori0317.gamemenumodoption;
+
+import net.morimori0317.gamemenumodoption.explatform.GMMOExpectPlatform;
+
+public class GameMenuModOption {
+    public static final String MOD_ID = "gamemenumodoption";
+
+    public static ClientConfig getConfig() {
+        return GMMOExpectPlatform.getConfig();
+    }
+}