+++ common/src/main/java/net/morimori0317/gamemenumodoption/WrappedTitleScreenModUpdateIndicator.java
@@ -1,0 +1,12 @@
+package net.morimori0317.gamemenumodoption;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.GuiGraphics;
+
+public interface WrappedTitleScreenModUpdateIndicator {
+    void resize(Minecraft minecraft, int width, int height);
+
+    void init();
+
+    void render(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick);
+}