+++ src/main/java/net/morimori0317/gmmo/GMMOPauseScreen.java
@@ -1,9 +1,0 @@
-package net.morimori0317.gmmo;
-
-import net.minecraft.client.gui.components.Button;
-
-public interface GMMOPauseScreen {
-    Button getModOptionButton();
-
-    boolean isShowPauseMenu();
-}